pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "manus-machina"
version = "5.0.0"
description = "Enterprise-Grade Multi-Agent Orchestration Framework - Complete Edition"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "João Campista", email = "joao@example.com"}
]
keywords = [
    "ai", "agents", "multi-agent", "llm", "framework",
    "sessions", "state-management", "evaluation", "memory",
    "clean-architecture", "ddd", "enterprise"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pydantic>=2.0.0",
    "httpx>=0.25.0",
    "structlog>=23.0.0",
]

[project.optional-dependencies]
# LLM Providers
llm = [
    "litellm>=1.0.0",
    "google-generativeai>=0.3.0",
    "openai>=1.0.0",
    "anthropic>=0.7.0",
]

# Database
database = [
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "psycopg2-binary>=2.9.0",  # PostgreSQL
    "pymysql>=1.1.0",  # MySQL
]

# Vector Stores & Memory
memory = [
    "faiss-cpu>=1.7.4",
    "pinecone-client>=2.2.0",
    "weaviate-client>=3.24.0",
    "chromadb>=0.4.0",
    "qdrant-client>=1.6.0",
]

# Evaluation
evaluation = [
    "rouge-score>=0.1.2",
]

# Observability
observability = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "prometheus-client>=0.18.0",
]

# Development
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

# Documentation
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.4.0",
    "mkdocstrings[python]>=0.23.0",
]

# All features
all = [
    "manus-machina[llm,database,memory,evaluation,observability]",
]

# All including dev
full = [
    "manus-machina[all,dev,docs]",
]

[project.urls]
Homepage = "https://github.com/JoaoCampista/manus-machina"
Documentation = "https://joaocampista.github.io/manus-machina"
Repository = "https://github.com/JoaoCampista/manus-machina"
Issues = "https://github.com/JoaoCampista/manus-machina/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["manus_machina*", "infrastructure*"]
exclude = ["tests*", "docs*", "examples*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

